Bookmark and Vote controllers working

+++ app/controllers/application_controller.rb
@@ -25,39 +25,46 @@ class ApplicationController < ActionController::Base
     !!current_user
   end
 
-  def bookmark(type, add)
+  # add is true for creating bookmark, false if deleting
+  def make_bookmark(type, add)
     if add
       @bookmark = current_user.bookmarks.new({ bookmarkable_type: type, bookmarkable_id: params[:id] })
       if @bookmark.save
-        # return current user to update user from store
-        render 'api/users/user', user: current_user
+        # return to controller and return updated resource
+        return true
       else
-        render json @bookmark.errors.full_messages, status: :unprocessable_entity
+        return :json => @bookmark.errors.full_messages, status: :unprocessable_entity
       end
     else
       @bookmark = current_user.bookmarks.where({ bookmarkable_type: type, bookmarkable_id: params[:id] }).first
+      if @bookmark.nil? 
+        return :json => ['No bookmark found'], status: :not_found
+      end 
       if @bookmark.destroy
-        render 'api/users/user', user: current_user
+        return true
       else
-        render json @bookmark.errors.full_messages, status: :unprocessable_entity
+        return :json => @bookmark.errors.full_messages, status: :unprocessable_entity
       end
     end
   end
 
-  def vote(type, add)
+  def make_vote(type, add)
     if add
       @vote = current_user.votes.new({ voteable_type: type, voteable_id: params[:id] })
       if @vote.save
-        render 'api/users/user', user: current_user
+        return true
       else
-        render json @vote.errors.full_messages, status: :unprocessable_entity
+        return :json => @vote.errors.full_messages, status: :unprocessable_entity
       end
     else
       @vote = current_user.votes.where({ voteable_type: type, voteable_id: params[:id] }).first
+      if @vote.nil?
+        return :json => ['No vote found'], status: :not_found
+      end
       if @vote.destroy
-        render 'api/users/user', user: current_user
+        return true
       else
-        render json @vote.errors.full_messages, status: :unprocessable_entity
+        return :json => @vote.errors.full_messages, status: :unprocessable_entity
       end
     end
   end
